# openapi.json
{
  "components": {
    "schemas": {
...
      "MoneyRequest": {
        "examples": [
          {
            "amount": "100.00",
            "currency": "EUR",
            "description": "salary"
          }
        ],
        "properties": {
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,18}|(?=[\\d.]{1,21}0*$)\\d{0,18}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "description": "Positive, at most two decimal places. Sent as a JSON string so no float rounding happens in transit.",
            "examples": [
              "100.00"
            ],
            "title": "Amount"
          },
          "currency": {
            "description": "Must match the wallet's currency; a mismatch is refused, not converted.",
            "examples": [
              "EUR"
            ],
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "description": {
            "default": "",
            "description": "Free text recorded on the ledger entry. Part of the idempotency fingerprint: the same key with a different description is a different request.",
            "examples": [
              "salary"
            ],
            "maxLength": 200,
            "title": "Description",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency"
        ],
        "title": "MoneyRequest",
        "type": "object"
      },
...
      "OpenWalletRequest": {
        "examples": [
          {
            "currency": "EUR"
          }
        ],
        "properties": {
          "currency": {
            "description": "ISO 4217 alphabetic code, uppercase. Fixed for the wallet's life.",
            "examples": [
              "EUR"
            ],
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "currency"
        ],
        "title": "OpenWalletRequest",
        "type": "object"
      },
      "Problem": {
        "description": "RFC 9457 `application/problem+json`.",
        "examples": [
          {
            "code": "WALLET_INSUFFICIENT_FUNDS",
            "correlation_id": "489b1a95-1123-43ef-b6f6-03bbb5e15248",
            "details": {
              "balance": "70.00",
              "requested": "1000.00"
            },
            "instance": "/v1/wallets/01a0084a-d15c-7122-bc8c-a81b38e8a2f4/withdrawals",
            "status": 409,
            "title": "Wallet insufficient funds",
            "type": "about:blank#wallet_insufficient_funds"
          }
        ],
        "properties": {
          "code": {
            "description": "Machine-readable code. Branch on this, never on `title` — titles are prose and may be reworded; codes are part of the contract.",
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Quote this when reporting the failure; it is the key to the server-side log line. Present on every response this API produces.",
            "title": "Correlation Id"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured context, shape depending on `code`. Never contains the rejected value or any server internals.",
            "title": "Details"
          },
          "instance": {
            "description": "Path of the request that produced the problem.",
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code, repeated for clients that lose it.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short human-readable summary. Safe to display.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI reference identifying the problem kind.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "instance"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "WalletPageResponse": {
        "properties": {
          "items": {
            "description": "Newest first.",
            "items": {
              "$ref": "#/components/schemas/WalletResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` for the next page. `null` means this was the last one. Opaque: it encodes a position, not an offset, so rows inserted while you page are neither skipped nor repeated.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "WalletPageResponse",
        "type": "object"
      },
      "WalletResponse": {
        "properties": {
          "balance": {
            "description": "Decimal string, never a JSON number.",
            "examples": [
              "100.00"
            ],
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance",
            "type": "string"
          },
          "currency": {
            "description": "ISO 4217.",
            "examples": [
              "EUR"
            ],
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "description": "UUIDv7, so id order is creation order.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_id": {
            "description": "The authenticated caller. Always yours.",
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "status": {
            "description": "`open` or `closed`. A closed wallet refuses movements.",
            "examples": [
              "open"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "description": "Optimistic concurrency counter, incremented by every movement. Exposed because it is the number that explains a 409.",
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "owner_id",
          "balance",
          "currency",
          "status",
          "version"
        ],
        "title": "WalletResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "Owner bearer token": {
        "bearerFormat": "UUID",
        "description": "Any UUID, and it is the owner id — there is no signup, no login endpoint and no token service in this template.\n\nThe bundled `DevIdentityProvider` treats the bearer token as the identity, so `Authorization: Bearer 018f4b2c-...` opens a fresh account and `uuidgen` is the whole onboarding flow. Wallets are scoped to it: use the same value again to see them, a different one to be somebody else.\n\nThat adapter refuses to start outside a local run, which is enforced by the boot guard. A real deployment swaps one provider method for a JWT/JWKS adapter, and this description changes with it.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "title": "evenkeel",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/wallets": {
      "get": {
        "description": "Cursor-paginated, newest first.\n\nPass `next_cursor` back as `cursor` to walk further. Offsets are not\nsupported on purpose: they make deep pages linearly slower and let\nconcurrent inserts shift rows between pages.",
        "operationId": "list_wallets_v1_wallets_get",
        "parameters": [
          {
            "description": "Page size. The maximum is 100 — a page nobody can request is a page nobody can use to scan the table.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Page size. The maximum is 100 — a page nobody can request is a page nobody can use to scan the table.",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque, from `next_cursor` on the previous page. Absent means the first page; `next_cursor: null` in a response means the last. Cursors are not offsets: they do not skip or repeat rows when something is inserted while you page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque, from `next_cursor` on the previous page. Absent means the first page; `next_cursor: null` in a response means the last. Cursors are not offsets: they do not skip or repeat rows when something is inserted while you page.",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "No credential was supplied, or it was not a valid owner id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "The request failed validation, or violated a domain rule that is not a conflict — mixing currencies, for example. `details.errors` names the offending fields but never echoes their values."
          }
        },
        "security": [
          {
            "Owner bearer token": []
          }
        ],
        "summary": "List the caller's wallets",
        "tags": [
          "wallets"
        ]
      },
      "post": {
        "description": "Handlers stay this thin.\n\nParse, call one use case, shape the response. No business rules, no\nrepository access, no transaction handling -- that is what keeps the same\nuse case reusable from a worker, a CLI or an MCP tool.",
        "operationId": "open_wallet_v1_wallets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenWalletRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "No credential was supplied, or it was not a valid owner id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "The request failed validation, or violated a domain rule that is not a conflict — mixing currencies, for example. `details.errors` names the offending fields but never echoes their values."
          }
        },
        "security": [
          {
            "Owner bearer token": []
          }
        ],
        "summary": "Open a wallet",
        "tags": [
          "wallets"
        ]
      }
    },
    "/v1/wallets/{wallet_id}": {
      "get": {
        "description": "A wallet owned by someone else is reported as absent, not forbidden.",
        "operationId": "get_wallet_v1_wallets__wallet_id__get",
        "parameters": [
          {
            "description": "A wallet belonging to the authenticated owner. One that belongs to somebody else is reported as absent, so this cannot be used to discover which ids exist.",
            "in": "path",
            "name": "wallet_id",
            "required": true,
            "schema": {
              "description": "A wallet belonging to the authenticated owner. One that belongs to somebody else is reported as absent, so this cannot be used to discover which ids exist.",
              "format": "uuid",
              "title": "Wallet Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "No credential was supplied, or it was not a valid owner id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "No such wallet for this owner. A wallet that exists but belongs to someone else is reported as absent, so the status cannot be used to probe which ids are real."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "The request failed validation, or violated a domain rule that is not a conflict — mixing currencies, for example. `details.errors` names the offending fields but never echoes their values."
          }
        },
        "security": [
          {
            "Owner bearer token": []
          }
        ],
        "summary": "Read one wallet",
        "tags": [
          "wallets"
        ]
      }
    },
    "/v1/wallets/{wallet_id}/deposits": {
      "post": {
        "description": "Send `Idempotency-Key` and a retry cannot apply twice.\n\nThe response then carries `replayed: true` and the original `entry_id`.\nReusing a key with a different payload is refused with 409 rather than\nsilently confirmed.",
        "operationId": "deposit_v1_wallets__wallet_id__deposits_post",
        "parameters": [
          {
            "description": "A wallet belonging to the authenticated owner. One that belongs to somebody else is reported as absent, so this cannot be used to discover which ids exist.",
            "in": "path",
            "name": "wallet_id",
            "required": true,
            "schema": {
              "description": "A wallet belonging to the authenticated owner. One that belongs to somebody else is reported as absent, so this cannot be used to discover which ids exist.",
              "format": "uuid",
              "title": "Wallet Id",
              "type": "string"
            }
          },
          {
            "description": "Reuse the same value when retrying a request you are unsure completed. The movement is applied once and the retry returns the original entry with `replayed: true`; a new value moves money again.\n\nScoped per owner, so two clients picking `1` do not collide. Reusing one with a different payload is refused with `409` rather than confirming an operation that never ran, and a duplicate that arrives while the first is still running is refused too — there is no result to replay yet.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Reuse the same value when retrying a request you are unsure completed. The movement is applied once and the retry returns the original entry with `replayed: true`; a new value moves money again.\n\nScoped per owner, so two clients picking `1` do not collide. Reusing one with a different payload is refused with `409` rather than confirming an operation that never ran, and a duplicate that arrives while the first is still running is refused too — there is no result to replay yet.",
              "examples": [
                "7d1f8e0a-4b3c-4e2a-9f1d-2c5b6a7e8d90"
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoneyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "No credential was supplied, or it was not a valid owner id."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Risk assessment refused the movement. Nothing changed, and retrying with the same payload will be refused again. `details.reference` is the provider's id for the decision, for a support conversation."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "No such wallet for this owner. A wallet that exists but belongs to someone else is reported as absent, so the status cannot be used to probe which ids are real."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "The movement was refused and nothing changed: insufficient funds, the wallet is closed, another writer holds it, a concurrent write won, an earlier request with this idempotency key is still running, or the key was reused with a different payload. `code` distinguishes them."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "The request failed validation, or violated a domain rule that is not a conflict — mixing currencies, for example. `details.errors` names the offending fields but never echoes their values."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Too many movements for this owner. `details.retry_after_seconds` says how long to wait. Nothing was applied."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "A dependency the operation needs is unreachable. Transient; retry with the same idempotency key."
          }
        },
        "security": [
          {
            "Owner bearer token": []
          }
        ],
        "summary": "Credit a wallet",
        "tags": [
          "wallets"
        ]
      }
    },
    "/v1/wallets/{wallet_id}/entries": {
      "get": {
        "description": "Append-only history for one wallet, newest first.\n\nThe balance is a cache of these entries, so replaying them reconstructs it.",
        "operationId": "list_entries_v1_wallets__wallet_id__entries_get",
        "parameters": [
          {
            "description": "A wallet belonging to the authenticated owner. One that belongs to somebody else is reported as absent, so this cannot be used to discover which ids exist.",
            "in": "path",
            "name": "wallet_id",
            "required": true,
            "schema": {
              "description": "A wallet belonging to the authenticated owner. One that belongs to somebody else is reported as absent, so this cannot be used to discover which ids exist.",
              "format": "uuid",
              "title": "Wallet Id",
              "type": "string"
            }
          },
          {
            "description": "Page size. The maximum is 100 — a page nobody can request is a page nobody can use to scan the table.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Page size. The maximum is 100 — a page nobody can request is a page nobody can use to scan the table.",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque, from `next_cursor` on the previous page. Absent means the first page; `next_cursor: null` in a response means the last. Cursors are not offsets: they do not skip or repeat rows when something is inserted while you page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque, from `next_cursor` on the previous page. Absent means the first page; `next_cursor: null` in a response means the last. Cursors are not offsets: they do not skip or repeat rows when something is inserted while you page.",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "No credential was supplied, or it was not a valid owner id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "No such wallet for this owner. A wallet that exists but belongs to someone else is reported as absent, so the status cannot be used to probe which ids are real."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "The request failed validation, or violated a domain rule that is not a conflict — mixing currencies, for example. `details.errors` names the offending fields but never echoes their values."
          }
        },
        "security": [
          {
            "Owner bearer token": []
          }
        ],
        "summary": "List ledger entries for a wallet",
        "tags": [
          "wallets"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "docker compose up",
      "url": "http://localhost:8000"
    },
    {
      "description": "This deployment",
      "url": "/"
    }
  ],
...
}